requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
evdev==1.7.1
idna==3.10
keyboard==0.13.5
numpy==2.0.0
PyAudio==0.2.14
pycparser==2.22
pynput==1.7.7
python-xlib==0.33
requests==2.32.3
six==1.17.0
sounddevice==0.5.1
soundfile==0.12.1
urllib3==2.2.3

pip~=24.0
Jinja2~=3.1.4
filelock~=3.16.1
python3-xlib~=0.15
setuptools~=75.6.0
numba~=0.60.0
typing_extensions~=4.12.2
llvmlite~=0.43.0
regex~=2024.11.6
sympy~=1.13.1
mpmath~=1.3.0
torch~=2.5.1
tqdm~=4.67.1
triton~=3.1.0
fsspec~=2024.10.0
MarkupSafe~=3.0.2
more-itertools~=10.5.0
tiktoken~=0.8.0
networkx~=3.4.2
PyMsgBox~=1.0.9
pyperclip~=1.9.0
MouseInfo~=0.1.3
PyScreeze~=1.0.1
pytweening~=1.2.0
PyGetWindow~=0.0.9
PyAutoGUI~=0.9.54
PyRect~=0.2.0
openai-whisper~=20240930
soundcard
numpy
google-generativeai
